requirements.txt
streamlit
dotenv
numpy
pandas
openpyxl
xlsxwriter
plotly
requests
sqlite3
websocket-client
json
os
math
io
pathlib
logging
csv
python-dotenv
